sensor_portal/requirements.txt

Django==4.2
django-debug-toolbar==3.7.0
#django-dbbackup==4.0.2

# auth
bridgekeeper==0.9
pyjwt==2.9.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0

# REST
djangorestframework==3.15
djangorestframework-gis==1.0
django-filter==22.1

# Django extra fields
django-sizefield==2.1.0
django-timezone-field==7.1
django-encrypted-model-fields==0.6.5

# celery
celery==5.2.7
redis==4.3.3
hiredis==2.0.0
django-celery-beat==2.5.0
django-celery-results==2.5.1

# SSH access
paramiko==2.11.0
scp==0.14.4

# Other packages
pytz==2022.1
python-magic==0.4.27
numpy==1.26.4
pandas==1.4.2
plotly==5.9.0
django-crispy-forms==1.14.0
django-autocomplete-light==3.9.4
crispy-bootstrap5==0.7
drf-yasg==1.21.10
#django-recaptcha==3.0.0
#bagit==1.8.1
#scikit-image==0.19.3
#git+https://github.com/neutronX/django-markdownx@master
#pygments==2.12.0
#django-dbbackup==4.0.2
#thefuzz==0.20.0
thefuzz==0.20.0
requests==2.32.3

# testing
pytest==8.3.3
pytest-django==4.9.0
factory-boy==3.3.1

# problematic to install in my local env
psycopg2==2.9.3
Pillow==9.1.1

# archiving
bagit==1.8.1


